Update hashing code structure
