+++ ec/src/hashing/curve_maps/mod.rs
@@ -1,0 +1,1 @@
+pub mod shallue_van_de_woestijne;
